Serverside:
Added:
Ability to turn on/off your comm.

Improved:
Informational messages no longer display the unnecessary "Success" or "Error" as you can tell by the message's color.

Admins now always have cheataccess if they have the proper bitvalue.

You can now differ between all chat and regular chat. There an @ before the message if it's all chat. Currently you see both all chat and regular chat at the same time. (May be changed, may be a feature).

Fixed:
Possibly fixed a skill point bug.

Clientside:

Improved:
Upped the limit of cg_fov from 95 to 120.

Removed cheat protection from some cvars.

git-svn-id: http://openrp.googlecode.com/svn/trunk@217 3c299bc4-26c0-bb6f-2937-d8014291f781

+++ source/game/g_active.cpp
@@ -2730,7 +2730,7 @@ void ClientThink_real( gentity_t *ent ) {
 	//1.3 (DarthDie) -- update as soon as we have had skill changed, so we can do AddText
 	if(!ent->NPC && !(ent->r.svFlags & SVF_BOT) && client->skillUpdated /*&& client->skillDebounce < level.time*/)
 	{
-		trap_SendServerCommand(ent->s.number, va("nfr %i %i %i", (int) ent->client->sess.skillPoints, 0, ent->client->sess.sessionTeam));
+		trap_SendServerCommand(ent->s.number, va("nfr %i %i %i", ent->client->sess.skillPoints, 0, ent->client->sess.sessionTeam));
 		//client->skillDebounce = level.time + 5000;
 		client->skillUpdated = qfalse;
 	}
